// packages/datadog-api-client-v1/models/QueryValueWidgetRequest.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FormulaAndFunctionQueryDefinition } from './FormulaAndFunctionQueryDefinition';
import { FormulaAndFunctionResponseFormat } from './FormulaAndFunctionResponseFormat';
import { LogQueryDefinition } from './LogQueryDefinition';
import { ProcessQueryDefinition } from './ProcessQueryDefinition';
import { WidgetAggregator } from './WidgetAggregator';
import { WidgetConditionalFormat } from './WidgetConditionalFormat';
import { WidgetFormula } from './WidgetFormula';
/**
* Updated query value widget.
*/
export declare class QueryValueWidgetRequest {
    'aggregator'?: WidgetAggregator;
    'apmQuery'?: LogQueryDefinition;
    /**
    * List of conditional formats.
    */
    'conditionalFormats'?: Array<WidgetConditionalFormat>;
    'eventQuery'?: LogQueryDefinition;
    /**
    * List of formulas that operate on queries. **This feature is currently in beta.**
    */
    'formulas'?: Array<WidgetFormula>;
    'logQuery'?: LogQueryDefinition;
    'networkQuery'?: LogQueryDefinition;
    'processQuery'?: ProcessQueryDefinition;
    'profileMetricsQuery'?: LogQueryDefinition;
    /**
    * TODO.
    */
    'q'?: string;
    /**
    * List of queries that can be returned directly or used in formulas. **This feature is currently in beta.**
    */
    'queries'?: Array<FormulaAndFunctionQueryDefinition>;
    'responseFormat'?: FormulaAndFunctionResponseFormat;
    'rumQuery'?: LogQueryDefinition;
    'securityQuery'?: LogQueryDefinition;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): QueryValueWidgetRequest;
    static serialize(data: QueryValueWidgetRequest): {
        [key: string]: any;
    };
    constructor();
}